FIX-pis to biginteger

+++ database/migrations/2025_10_14_160814_create_points_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('points', function (Blueprint $table) {
             $table->id();
-            $table->integer('pis')->nullable();
+            $table->bigInteger('pis')->nullable();
             $table->date('date')->nullable();
             $table->time('time')->nullable();
             $table->string('type')->nullable();
